requirements.txt
PyYAML>=5.1
requests>=2.14.2
typing-extensions; python_version<='3.7'
